Adding Foundation Icon Fonts 3
Source: http://zurb.com/playground/foundation-icon-fonts-3 and
http://zurb.com/playground/uploads/upload/upload/288/foundation-icons.zi
p

CSS min’d, removed preview.html - other files as-is. Updated homepage
and keywords in package.json to bring up-to-date. Supercedes #2287,
#2288, #2289 and #2290.

Previous package.json did not validate due to semver error: 2.0 is not
a valid version number. This resolves it. Took me a while to figure it
out, but it’s now valid and 18% sexier.