fastclick-1.0.2: Strip UMD wrapper.
The UMD wrapper was inserted by the source of minification,
http://build.origami.ft.com/bundles/js?modules=fastclick . You can see
better the inserted UMD wrapper at that I have removed at
http://build.origami.ft.com/bundles/js?modules=fastclick&minify=none
. The UMD wrapper, which is not included in cdnjs’s `fastclick.js` for
fastclick-1.0.2 or in fastclick-1.0.0’s `fastclick.min.js`, prevents
the library from exporting the `FastClick` symbol into the global
namespace. It is only usable if used in a CommonJS environment and
forces developers to call `require()`. It is not how fastclick is
intended to be packaged and breaks most existing scripts which
directly use `FastClick`.